#TASK-003 - Sign In Page - Validations and refactoring
