+++ src/app/auth/auth.component.scss
@@ -74,18 +74,7 @@ input[type="submit"] {
   }
 }
 
-.sign-up-link{
-  margin: 20px;
-  font-size: 16px;
-  color: #a5a5a5;
-  cursor: pointer;
-}
 
-.sign-up-link__span{
-  color: #612F74;
-  font-weight: bold;
-  
-}
 
 .icon__email {
   background: url(https://img.icons8.com/ios/20/000000/user.png) no-repeat scroll 15px 14px;
