+++ src/app/auth/auth.module.ts
@@ -4,7 +4,6 @@ import { AuthRoutingModule } from './auth.routing.module';
 import { DirectiveModule } from '../directives/directive.module';
 import { AuthComponent } from './auth.component';
 import { SignInComponent } from './sign-in/sign-in.component';
-import { SignUpComponent } from './sign-up/sign-up.component';
 import { MaterialModule } from '../material/material.module';
 import { HeadTitleComponent } from './head-title/head-title.component';
 import { SocialMediaComponent } from './social-media/social-media.component';
@@ -14,7 +13,6 @@ import { FormsModule, ReactiveFormsModule } from '@angular/forms';
   declarations: [
     AuthComponent,
     SignInComponent,
-    SignUpComponent,
     HeadTitleComponent,
     SocialMediaComponent
   ],
