+++ src/app/auth/sign-in/sign-in.component.scss
@@ -2,3 +2,16 @@
     width: 50%;
     color: white;
 }
+
+.sign-up-link{
+    margin: 20px;
+    font-size: 16px;
+    color: #a5a5a5;
+    cursor: pointer;
+  }
+  
+  .sign-up-link__span{
+    color: #612F74;
+    font-weight: bold;
+    
+  }
