--- src/app/auth/sign-up/sign-up.component.html
@@ -1,50 +1,0 @@
-<form class="form" (ngSubmit)="onSignup()" [formGroup]="form">
-
-  <input 
-    type="text" 
-    id="email" 
-    name="email" 
-    placeholder="Email" 
-    formControlName="email" 
-    class="icon__email" 
-    [ngClass]="form.get('email').hasError('pattern') && form.get('email').touched 
-                || form.get('email').hasError('required') && form.get('email').touched 
-                ? 'input-error' : ''">
-
-  <input 
-    type="password" 
-    id="password" 
-    name="password" 
-    placeholder="Password" 
-    formControlName="password"
-    class="icon__password" 
-    minlength="6" 
-    [ngClass]="form.get('password').hasError('minlength') && form.get('password').touched 
-               || form.get('password').hasError('required') && form.get('password').touched  
-               ? 'input-error' : ''">
-
-  <input 
-    type="password" 
-    id="repassword" 
-    name="repassword" placeholder="Retype passowrd" 
-    formControlName="repassword"
-    class="icon__password" 
-    minlength="6"
-    [ngClass]="form.get('repassword').hasError('minlength') && form.get('repassword').touched 
-               || form.get('repassword').hasError('required') && form.get('repassword').touched  
-               ? 'input-error' : ''">
- 
-    <input 
-      type="submit" 
-      value="SIGN UP" 
-      [disabled]="form.get('email').hasError('required') ||
-                  form.get('email').hasError('pattern') ||
-                  form.get('password').hasError('minlength') || 
-                  form.get('password').hasError('required') || 
-                  form.get('repassword').hasError('minlength') || 
-                  form.get('repassword').hasError('required') || 
-                  passwordsDontMatch()"
-                  >
-  <br>
-  <small *ngIf="passwordDontMatch" class="match-password">Passwords don't match</small>
-</form>
