--- src/app/auth/sign-up/sign-up.component.scss
@@ -1,3 +1,0 @@
-.match-password{
-    font-weight: bold;
-}
